AL: prevent duplicate appends

+++ flake.nix
@@ -77,13 +77,21 @@
 
                 [[ -f justfile  ]] && command -v just >/dev/null 2>&1 && just --list --unsorted
 
-                cp justfile $HOME
-                python3 -m venv $HOME/.devenv && \
-                source "$HOME/.devenv/bin/activate" && \
-                pip install --index-url=https://pypi.org/simple poetry
+                if [[ ! -e "$HOME/justfile" ]]; then
+                  cp justfile $HOME
+                fi
 
-                echo 'source "$HOME/.devenv/bin/activate"' | tee -a $HOME/.zshrc > /dev/null
-                echo 'nix develop "github:scorbettum:local-dev"' | tee -a $HOME/.zhsrc > /dev/null
+                if [[ ! grep -Fxq 'source "$HOME/.devenv/bin/activate"' $HOME/.zshrc ]]; then
+                  python3 -m venv $HOME/.devenv && \
+                  source "$HOME/.devenv/bin/activate" && \
+                  pip install --index-url=https://pypi.org/simple poetry
+
+                  echo 'source "$HOME/.devenv/bin/activate"' | tee -a $HOME/.zshrc > /dev/null
+                fi
+
+                if [[ ! grep -Fxq 'nix develop "github:scorbettum:local-dev"' $HOME/.zshrc   ]]; then
+                  echo 'nix develop "github:scorbettum:local-dev"' | tee -a $HOME/.zhsrc > /dev/null
+                fi
               '';
           };
         }) // {
